requirements.txt
streamlit
langchain-openai>=0.1.0
langchain>=0.1.0
openai
python-dotenv
